move serialization logic out of virtu and into bizhawk, this was our shenanigans
